feat: add cover image upload hook
